wip - linux usb gadget uvc video camera
I think this is making the computer into a uvc webcam, I was hoping to get it to recognize an endoscope that works when plugged into an android phone (with a proprietary app)
I think the keyword is OTG (on the go). Which is usually an adapter that goes between an android and a regular USB device. But in the case of the endoscope it's built in so you can plug the webcam into an android.